aster_mlsdisk

Trait BlockSet

pub trait BlockSet: Sync + Send {
    // Required methods
    fn read(&self, pos: BlockId, buf: BufMut<'_>) -> Result<(), Error>;
    fn write(&self, pos: BlockId, buf: BufRef<'_>) -> Result<(), Error>;
    fn subset(&self, range: Range<BlockId>) -> Result<Self, Error>
       where Self: Sized;
    fn flush(&self) -> Result<(), Error>;
    fn nblocks(&self) -> usize;

    // Provided methods
    fn read_slice(&self, offset: usize, buf: &mut [u8]) -> Result<(), Error> { ... }
    fn write_slice(&self, offset: usize, buf: &[u8]) -> Result<(), Error> { ... }
}

A fixed set of data blocks that can support random reads and writes.

Thread safety

BlockSet is a data structure of interior mutability. It is ok to perform I/O on a BlockSet concurrently in multiple threads. BlockSet promises the atomicity of reading and writing individual blocks.

Required Methods

fn read(&self, pos: BlockId, buf: BufMut<'_>) -> Result<(), Error>

Read one or multiple blocks at a specified position.

fn write(&self, pos: BlockId, buf: BufRef<'_>) -> Result<(), Error>

Write one or multiple blocks at a specified position.

fn subset(&self, range: Range<BlockId>) -> Result<Self, Error>
where Self: Sized,

Get a subset of the blocks in the block set.

fn flush(&self) -> Result<(), Error>

Ensure that blocks are persisted to the disk.

fn nblocks(&self) -> usize

Returns the number of blocks.

Provided Methods

fn read_slice(&self, offset: usize, buf: &mut [u8]) -> Result<(), Error>

Read a slice of bytes at a specified byte offset.

fn write_slice(&self, offset: usize, buf: &[u8]) -> Result<(), Error>

Write a slice of bytes at a specified byte offset.